requirements.txt
requests
beautifulsoup4
feedparser
schedule
python-dotenv
pyyaml
lxml
youtube-search-python
